Fix compilation with boost 1.60
Newer boost versions use complicated type traits for boost::next and
boost::prior, which do not work with the RandomAccessList iterators.
The long term solution is to use std::next and std::prev, for now supply
simple replacements for compilers that do not support C++11 yet.

This is was cherry picked from b5963300 and 7e72c1d0d3.