Revert "toc bind: 'dialog-show' -> 'dialog-toggle' (#8388)"
This reverts commit bede4d320bfff26fcbea396b516b92eb9595940c.

As JMarc notes [1], if we replaced show with toggle, we would lose some
functionality:

  There is one use that is not possible with this change (as I argued
  before), which is a fully keyboard-based navigation.

  With LyX 2.1, I can do C-M-o and navigate with cursor in the
  outline. I can do the same with advanced S&R. IN this sense,
  toggling is a step back IMO.

The plan now is to find another way to close the dialog (without
using toggle), such as having ESC close it or another key bind.

Ticket #8388 will stay closed because 'show' and 'toggle' are too
close for the cost of having two separate bindings.

[1] http://www.mail-archive.com/lyx-devel@lists.lyx.org/msg187693.html