Fix several issues with instant previews.
This is a collection of changes to the python scripts in master made
for trying to sort out the mess that had accumulated about previews.
After this audit, it turns out that bugs #6369, #9354, #9508, and #9510
are now fixed. If pdftocairo and epstopdf are vailable, it also makes
the conversion performed by the legacy conversion paths much faster
and improves the quality of the generated bitmaps.