Revert "Update previews after pasting (#6211)"
As JMarc notes on the ML, the referenced commit modifies a full
buffer just in the middle of a dispatch action.

Further, this commit lead to an assertion [1], although it is not
clear to me whether the root cause of the assertion is the commit
that is being reverted here, or one of the issues we have with
previews and macros.

This reverts commit 6ac04e21f46731f447fc53c184c7193f68ed3ee0.

[1] http://www.mail-archive.com/lyx-devel@lists.lyx.org/msg187921.html