Ensure the Greek Intro.lyx compiles with non-TeX fonts.
Prevent loading of lmodern.sty if non-TeX fonts are set and
Define DejaVu as non-TeX font.

This ensures that non-TeX fonts are used if requested (lmodern.sty set
the 8-bit version of LatinModern if compiling with LuaTeX).
Also ensure that a font with all characters is used. (LatinModern misses small
Greek characters.)

Similar actions for other manuals and examples with Greek or Cyrillic script
will solve some more export tests with non-TeX fonts.