Fix inputenc for some templates.
Some templates had the inputenc setting "default" (meaning 
use the language default encoding, (not TeX's default which is 7-bit ASCII) but
don't load inputenc, leave handling of non-ASCII characters to the author). 

This is no apparent problem as long as no non-ASCII character is used in the 
document. It is very unsave, however, as any non-ASCII character used in the 
document has the potential to either make compilation fail or be
misrepresented in the output.